add a removed_item attribute to help set prediction
some items are consumable or can be removed. e.g. airballoon, berries, boosterenergy. Before this change the bot would allow any item in a set, but this narrows it down to the removed item. A removed item can only be set once on a pkmn. This logic is not perfect but it should be strictly better than before