deal with transform on bot's side
the transform() fn is some old ass code from before I had the battle_modifier module parsing both sides. This changes it to parse both sides when a pkmn is transforming. Transform is still not handled well by the engine and the bot will make some bad plays when transform/imposter are in play for either side but this should make it so that it at least does not crash